fix compiler warning
(cherry picked from commit d4761549c1c0c0e7353d34279fd44b1fb7a8dce5)

+++ src/cgutils.cpp
@@ -933,7 +933,7 @@ static Value *mark_julia_type(Value *v, jl_value_t *jt)
 static bool deserves_sret(jl_value_t *dt, Type *T)
 {
     assert(jl_is_datatype(dt));
-    return jl_datatype_size(dt) > sizeof(void*) && !T->isFloatingPointTy();
+    return (size_t)jl_datatype_size(dt) > sizeof(void*) && !T->isFloatingPointTy();
 }
 
 // --- generating various field accessors ---
